+++ transfer_handler.go
@@ -1,30 +1,33 @@
 package main
 
 import (
+	"context"
 	"fmt"
+	alephium "github.com/alephium/go-sdk"
 	"github.com/sirupsen/logrus"
-	"github.com/touilleio/alephium-go-client"
+	"sync"
 	"time"
 )
 
 type transferHandler struct {
-	alephiumClient     *alephium.Client
+	alephiumClient     *alephium.APIClient
 	walletName         string
 	walletPassword     string
 	mnemonicPassphrase string
 	transferAddress    string
-	transferMinAmount  alephium.ALPH
+	transferMinAmount  ALPH
 	transferFrequency  time.Duration
 	immediate          bool
 	metrics            *metrics
 	log                *logrus.Logger
+	concurrentExecLock *sync.RWMutex
 }
 
-func newTransferHandler(alephiumClient *alephium.Client, walletName string, walletPassword string,
+func newTransferHandler(alephiumClient *alephium.APIClient, walletName string, walletPassword string,
 	mnemonicPassphrase string, transferAddress string, transferMinAmount string, transferFrequency time.Duration,
 	immediate bool, metrics *metrics, log *logrus.Logger) (*transferHandler, error) {
 
-	minAlf, ok := alephium.ALPHFromCoinString(transferMinAmount)
+	minAlf, ok := ALPHFromCoinString(transferMinAmount)
 	if !ok {
 		return nil, fmt.Errorf("transferMinAmount %s is not a valid ALPH transfer amoount", transferMinAmount)
 	}
@@ -40,21 +43,22 @@ func newTransferHandler(alephiumClient *alephium.Client, walletName string, wall
 		immediate:          immediate,
 		metrics:            metrics,
 		log:                log,
+		concurrentExecLock: &sync.RWMutex{},
 	}
 
 	return handler, nil
 }
 
-func (h *transferHandler) handle() error {
+func (h *transferHandler) handle(ctx context.Context, log *logrus.Entry) error {
 	if h.immediate {
-		err := h.transfer()
+		err := h.transfer(ctx, log)
 		if err != nil {
 			h.log.Debugf("Got an error while immediately transferring some amount. Err = %v", err)
 			return err
 		}
 	}
 	for range time.Tick(h.transferFrequency) {
-		err := h.transfer()
+		err := h.transfer(ctx, log)
 		if err != nil {
 			h.log.Debugf("Got an error while transferring some amount. Err = %v", err)
 			return err
@@ -63,71 +67,59 @@ func (h *transferHandler) handle() error {
 	return nil
 }
 
-func (h *transferHandler) transfer() error {
+func (h *transferHandler) transfer(ctx context.Context, log *logrus.Entry) error {
+
+	locked := h.concurrentExecLock.TryLock()
+	if !locked {
+		log.Warnf("Another transfer process is still running...")
+		return nil
+	}
+	defer h.concurrentExecLock.Unlock()
 
 	h.metrics.transferRun.Inc()
 
-	wallet, err := h.alephiumClient.GetWalletStatus(h.walletName)
+	wallet, err := getWalletStatus(ctx, h.alephiumClient, h.walletName, log)
 	if err != nil {
-		h.log.Debugf("Got an error calling wallet status, err = %v", err)
+		h.log.WithError(err).Debugf("Got an error calling wallet status, err = %v", err)
 		return err
 	}
 	if wallet.Locked {
-		_, err := h.alephiumClient.UnlockWallet(wallet.Name, h.walletPassword, h.mnemonicPassphrase)
+		err := unlockWallet(ctx, h.alephiumClient, wallet.WalletName, h.walletPassword, h.mnemonicPassphrase, log)
 		if err != nil {
-			h.log.Debugf("Got an error calling wallet unlock. Err = %v", err)
+			h.log.WithError(err).Debugf("Got an error calling wallet unlock. Err = %v", err)
 			return err
 		}
 	}
 
-	walletAddresses, err := h.alephiumClient.GetWalletAddresses(wallet.Name)
+	sweep := alephium.NewSweep(h.transferAddress)
+	sweepAllReq := h.alephiumClient.WalletsApi.PostWalletsWalletNameSweepAllAddresses(ctx, wallet.WalletName).Sweep(*sweep)
+	transferRes, _, err := sweepAllReq.Execute()
 	if err != nil {
-		h.log.Debugf("Got an error while getting wallet addresses. Err = %v", err)
+		h.log.WithError(err).Debugf("Got an error while sweeping all")
 		return err
 	}
 
-	for _, address := range getAddressesInRandomOrder(walletAddresses) {
+	for _, tx := range transferRes.GetResults() {
+		h.log.Infof("New tx %s,%d->%d just submitted", tx.TxId, tx.FromGroup, tx.ToGroup)
+		var txConfirmed *alephium.Confirmed
+		for txConfirmed == nil {
 
-		ok, err := h.alephiumClient.ChangeActiveAddress(wallet.Name, address)
-		if err != nil {
-			h.log.Debugf("Got an error calling change active address. Err = %v", err)
-			return err
+			txStatusReq := h.alephiumClient.TransactionsApi.GetTransactionsStatus(ctx)
+			txStatus, _, err := txStatusReq.TxId(tx.TxId).ToGroup(tx.FromGroup).ToGroup(tx.ToGroup).Execute()
+			if err != nil {
+				h.log.WithError(err).Debugf("Got an error while getting tx status for tx %s", tx.TxId)
+				return err
+			}
+			txConfirmed = txStatus.Confirmed
 		}
-		if !ok {
-			h.log.Warnf("Got a false while calling change active address. Not sure what this means yet...")
-		}
-
-		addressBalance, err := h.alephiumClient.GetAddressBalance(address, -1)
-		if err != nil {
-			h.log.Debugf("Got an error while getting address balance. Err = %v", err)
-			return err
-		}
-
-		amount := addressBalance.Balance.Subtract(addressBalance.LockedBalance)
-		h.log.Debugf("Address %s has %s available (i.e. not locked)", address, amount.PrettyString())
-		if amount.Amount == nil {
-			continue
-		}
-		if amount.Cmp(h.transferMinAmount) < 0 {
-			h.log.Debugf("Available amount %s is not above the threshold %s", amount.PrettyString(), h.transferMinAmount.PrettyString())
-			continue
-		}
-
-		h.log.Debugf("Will transfer all available (not locked) amount ~%s from address %s to %s", amount.PrettyString(), address, h.transferAddress)
-
-		tx, err := h.alephiumClient.SweepAll(wallet.Name, h.transferAddress)
-		if err != nil {
-			h.log.Debugf("Got an error calling transfer. Err = %v", err)
-			return err
-		}
-		h.metrics.txAmount.Add(amount.FloatALPH())
-		h.log.Infof("New tx %s,%d->%d of ~%s from %s to %s", tx.TransactionId, tx.FromGroup, tx.ToGroup,
-			amount.PrettyString(), address, h.transferAddress)
+		h.log.Infof("New tx %s,%d->%d is now inculded in block %s!", tx.TxId, tx.FromGroup, tx.ToGroup, txConfirmed.BlockHash)
+		// TODO: get block to find out the exact amount transferred to update counter consistently
 	}
+
 	return nil
 }
 
-func roundAmount(amount alephium.ALPH, txMinAmount alephium.ALPH, txMaxAmount alephium.ALPH) alephium.ALPH {
+func roundAmount(amount ALPH, txMinAmount ALPH, txMaxAmount ALPH) ALPH {
 	twiceMinAmount := txMinAmount.Multiply(2)
 	if amount.Cmp(twiceMinAmount) >= 0 {
 		txAmount := amount.Subtract(txMinAmount)
@@ -137,5 +129,5 @@ func roundAmount(amount alephium.ALPH, txMinAmount alephium.ALPH, txMaxAmount al
 			return txAmount
 		}
 	}
-	return alephium.ALPH{}
+	return ALPH{}
 }